requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-celery-results>=2.5

# Database
psycopg2-binary>=2.9
pgvector>=0.2

# Celery
celery>=5.3
redis>=5.0

# Whisper & Audio
openai-whisper>=20231117
librosa>=0.10
soundfile>=0.12
numpy>=1.24

# Embeddings
sentence-transformers>=2.2

# LoRA Fine-tuning
transformers>=4.35.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0
torch>=2.1.0

# Evaluation Metrics
jiwer>=3.0.0

# Model Export
ctranslate2>=3.20.0
faster-whisper>=0.9.0

# Utilities
python-dotenv>=1.0
gunicorn>=21.0
